Prepare for Guile-2.2
Guile-2.2 does not query (default-duplicate-binding-handler) at
module expand eval load time, unlike Guile-2.0 did.  As a
consequence, each individual module must declare its own
specified in the manual.

We, on purpose, systematically declare, in all modules importing
goops, the following:

  #:duplicates (merge-generics
  	        replace
  	        warn-override-core
	        warn
	        last)

This is because (a) we always want to merge-generics, (b) things
may and will change, such as imported module list, imported
modules lst order, ..., so even if it appears to be unnecessary
in some cases, then the next day it is, you forget and bang!

* g-golf/gi.scm:
* g-golf/gi/arg-info.scm:
* g-golf/gi/base-info.scm:
* g-golf/gi/enum-info.scm:
* g-golf/gi/glib.scm:
* g-golf/gi/gobject.scm:
* g-golf/gi/gobject/enum-flags.scm:
* g-golf/gi/gobject/type-info.scm:
* g-golf/gi/gobject/types.scm:
* g-golf/gi/object-info.scm:
* g-golf/gi/property-info.scm:
* g-golf/gi/type-info.scm:
* g-golf/gi/types.scm:
* g-golf/support.scm:
* g-golf/support/g-export.scm:  Updated as described here above.