Remove org files from the distro
G-Golf is part of GNU, and therefore will have a bug-g-golf mailing
list, once I create it, along with other ML.  So there is no real need
for a g-golf-bugs.org file.

The g-golf.org is/was supposed to be the project TODO file, but having
it under git and distributed sort of refrained me using it 'properly',
so I decided to maintain it locally.

* org/g-golf-bugs.org:
* org/g-golf.org:  Deleted.

* Makefile.am: Updated for the above.